+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/sidebar-layout.tsx
@@ -25,8 +25,6 @@ import {
 import {
   Book,
   FilePen,
-  Box,
-  CreditCard,
   Globe,
   KeyRound,
   LayoutTemplate,
@@ -42,7 +40,6 @@ import {
   User,
   Users,
   Webhook,
-  SquarePen,
 } from "lucide-react";
 import { useTheme } from "next-themes";
 import { usePathname } from "next/navigation";
@@ -243,10 +240,10 @@ const navigationItems: (Label | Item | Hidden)[] = [
     regex: /^\/projects\/[^\/]+\/email-themes\/[^\/]+$/,
     type: 'hidden',
   },
+  /*
   {
     name: "Payments",
     type: 'label',
-    requiresDevFeatureFlag: true,
   },
   {
     name: "Payments",
@@ -254,24 +251,7 @@ const navigationItems: (Label | Item | Hidden)[] = [
     regex: /^\/projects\/[^\/]+\/payments$/,
     icon: CreditCard,
     type: 'item',
-    requiresDevFeatureFlag: true,
-  },
-  {
-    name: "Offers",
-    href: "/payments/offers",
-    regex: /^\/projects\/[^\/]+\/payments\/offers$/,
-    icon: SquarePen,
-    type: 'item',
-    requiresDevFeatureFlag: true,
-  },
-  {
-    name: "Items",
-    href: "/payments/items",
-    regex: /^\/projects\/[^\/]+\/payments\/items$/,
-    icon: Box,
-    type: 'item',
-    requiresDevFeatureFlag: true,
-  },
+  },*/
   {
     name: "Configuration",
     type: 'label'
@@ -610,7 +590,7 @@ export default function SidebarLayout(props: { projectId: string, children?: Rea
         </div>
 
         {/* Content Body - Normal scrolling */}
-        <div className="flex-grow relative">
+        <div className="flex-grow relative flex flex-col">
           {props.children}
         </div>
       </div>