+++ packages/template/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @stackframe/stack
 
+## 2.8.36
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.36
+  - @stackframe/stack-sc@2.8.36
+  - @stackframe/stack-ui@2.8.36
+
+## 2.8.35
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.35
+  - @stackframe/stack-ui@2.8.35
+  - @stackframe/stack-sc@2.8.35
+
 ## 2.8.34
 
 ### Patch Changes